+++ testing/robot/Resources/vars.py
@@ -1,0 +1,18 @@
+##############################
+## Robot Framework Settings ##
+##############################
+RF_DEFAULT_TIMEOUT = 60
+RF_OPTIONAL_POPUP_TIMEOUT = 15
+RF_DEBUG_TRUE = True
+
+#########################
+## pytest-tui settings ##
+#########################
+REPO_BASE_DIR = "~/coding/pytest-tui/"
+
+#######################
+## Selenium Settings ##
+#######################
+BROWSER = "chrome"
+#BROWSER = "edge"
+#BROWSER = "firefox"